requirements.txt
anthropic
datasets
google-generativeai
openai
python-dotenv
requests
tqdm
